Encapsulated hooks logic
***
frontend
