+++ frontend/app/lib/components/board/boardStage.tsx
@@ -1,12 +1,14 @@
 'use client';
 
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { Stage } from 'react-konva';
 import GridLayer from '@/app/lib/components/board/gridLayer';
 import { Button } from '@heroui/button';
 import { BoxSelectIcon } from 'lucide-react';
 import { BoardElement } from '@/app/lib/types/definitions';
 import BoardLayer from '@/app/lib/components/board/boardLayer';
+import UseBoardZoom from '@/app/lib/hooks/useBoardZoom';
+import UseBoardNavigation from '@/app/lib/hooks/useBoardNavigation';
 
 const testBoardLayers = [
   [
@@ -77,9 +79,6 @@ const MAX_Y = 1000;
 const MIN_X = -2000;
 const MIN_Y = -2000;
 
-const MIN_SCALE = 0.5;
-const MAX_SCALE = 4;
-
 export const fitCoordinates = (
   x: number,
   y: number,
@@ -104,171 +103,13 @@ export const fitCoordinates = (
   return { x: newX, y: newY };
 };
 
-const useBoardNavigation = (stageRef, scale) => {
-  const [isDragging, setIsDragging] = useState(false);
-
-  useEffect(() => {
-    const stage: any = stageRef.current;
-    if (stage) {
-      // handling mouse navigation
-      const handleMouseDown = (e) => {
-        if (e.evt.button === 2) {
-          setIsDragging(true);
-        }
-      };
-      const handleMouseMove = (e) => {
-        if (isDragging) {
-          let newX = stage.x() + e.evt.movementX;
-          let newY = stage.y() + e.evt.movementY;
-
-          const fit = fitCoordinates(
-            newX,
-            newY,
-            window.innerWidth,
-            window.innerHeight,
-            scale,
-          );
-
-          stage.x(fit.x);
-          stage.y(fit.y);
-          stage.batchDraw();
-        }
-      };
-      const handleMouseUp = () => {
-        setIsDragging(false);
-      };
-      stage.on('mousedown', handleMouseDown);
-      stage.on('mousemove', handleMouseMove);
-      stage.on('mouseup', handleMouseUp);
-
-      // handling wheel navigation
-      const handleWheel = (e) => {
-        e.evt.preventDefault();
-        if (Math.abs(e.evt.deltaY as number) < 100) {
-          const newX = stage.x() - e.evt.deltaX;
-          const newY = stage.y() - e.evt.deltaY;
-
-          const fit = fitCoordinates(
-            newX,
-            newY,
-            window.innerWidth,
-            window.innerHeight,
-            scale,
-          );
-
-          stage.x(fit.x);
-          stage.y(fit.y);
-          stage.batchDraw();
-        }
-      };
-      stage.on('wheel', handleWheel);
-
-      // handling navigation with touch
-      const handleTouchStart = (e) => {
-        if (e.evt.touches.length === 2) {
-          setIsDragging(true);
-        }
-      };
-      const handleTouchMove = (e) => {
-        if (isDragging && e.evt.touches.length === 2) {
-          const touch1 = e.evt.touches[0];
-          const touch2 = e.evt.touches[1];
-          const movementX = (touch1.movementX + touch2.movementX) / 2;
-          const movementY = (touch1.movementY + touch2.movementY) / 2;
-
-          const newX = stage.x() + movementX;
-          const newY = stage.y() + movementY;
-
-          const fit = fitCoordinates(
-            newX,
-            newY,
-            window.innerWidth,
-            window.innerHeight,
-            scale,
-          );
-
-          stage.x(fit.x);
-          stage.y(fit.y);
-          stage.batchDraw();
-        }
-      };
-      const handleTouchEnd = () => {
-        setIsDragging(false);
-      };
-      stage.on('touchstart', handleTouchStart);
-      stage.on('touchmove', handleTouchMove);
-      stage.on('touchend', handleTouchEnd);
-
-      stage.container().addEventListener('contextmenu', (e) => {
-        e.preventDefault();
-      });
-
-      return () => {
-        stage.off('mousedown', handleMouseDown);
-        stage.off('mousemove', handleMouseMove);
-        stage.off('mouseup', handleMouseUp);
-
-        stage.off('wheel', handleWheel);
-
-        stage.off('touchstart', handleTouchStart);
-        stage.off('touchmove', handleTouchMove);
-        stage.off('touchend', handleTouchEnd);
-        stage
-          .container()
-          .removeEventListener('contextmenu', (e) => e.preventDefault());
-      };
-    }
-  }, [isDragging, scale]);
-};
-
-const useBoardZoom = (stageRef, scale, setScale) => {
-  useEffect(() => {
-    const stage = stageRef.current;
-    if (stage) {
-      const handleWheel = (e) => {
-        e.evt.preventDefault();
-
-        if (Math.floor(e.evt.deltaY as number) !== (e.evt.deltaY as number)) {
-          const scaleBy = 1.04;
-          const oldScale = stage.scaleX();
-          const pointer = stage.getPointerPosition();
-          const mousePointTo = {
-            x: (pointer.x - stage.x()) / oldScale,
-            y: (pointer.y - stage.y()) / oldScale,
-          };
-
-          const newScale =
-            e.evt.deltaY < 0 ? oldScale * scaleBy : oldScale / scaleBy;
-          if (!(MIN_SCALE >= newScale || MAX_SCALE <= newScale)) {
-            setScale(newScale);
-
-            stage.scale({ x: newScale, y: newScale });
-            const newPos = {
-              x: pointer.x - mousePointTo.x * newScale,
-              y: pointer.y - mousePointTo.y * newScale,
-            };
-            stage.position(newPos);
-            stage.batchDraw();
-          }
-        }
-      };
-
-      stage.on('wheel', handleWheel);
-
-      return () => {
-        stage.off('wheel', handleWheel);
-      };
-    }
-  }, [scale, setScale]);
-};
-
 export default function BoardStage() {
   const stageRef: any = useRef(null);
   const [layers, setLayers] = useState(testBoardLayers as BoardElement[][]);
   const [scale, setScale] = useState(1);
 
-  useBoardNavigation(stageRef, scale);
-  useBoardZoom(stageRef, scale, setScale);
+  UseBoardNavigation(stageRef, scale);
+  UseBoardZoom(stageRef, scale, setScale);
 
   const resetStage = () => {
     const stage = stageRef.current;
